Creating CRUD operation for users model
There is an intermediate file called ManagerUser that will help as intermediary between the user model and controller. This will help with the Exception handling process, and business logic isolation (referring to the model itself).

There is also some new packages like livewire because it was attempted to create the notification using this, but came up misunderstanding some things in the docs of livewire 3x so idecided to go simplier (template based using session, not component).

The routes for this part will come in the next commit due to the login development, I went straight into that after finishing the CRUD part.